Added proper student-facing README; fixed student-facing specs to not give warnings with rspec3
