+++ .rspec
@@ -1,2 +1,2 @@
---format spec
+--format documentation
 --color
