--- lib/part1.rb
@@ -1,11 +1,0 @@
-def sum arr
-  # YOUR CODE HERE
-end
-
-def max_2_sum arr
-  # YOUR CODE HERE
-end
-
-def sum_to_n? arr, n
-  # YOUR CODE HERE
-end
