--- lib/part2.rb
@@ -1,11 +1,0 @@
-def hello(name)
-  # YOUR CODE HERE
-end
-
-def starts_with_consonant? s
-  # YOUR CODE HERE
-end
-
-def binary_multiple_of_4? s
-  # YOUR CODE HERE
-end
