--- lib/part3.rb
@@ -1,3 +1,0 @@
-class BookInStock
-# YOUR CODE HERE
-end
