+++ spec/part1_spec.rb
@@ -1,69 +1,39 @@
+require 'ruby_intro.rb'
 describe 'Ruby intro part 1' do
   describe "#sum" do
-    it "should be defined" do
-      expect { sum([1,3,4]) }.not_to raise_error
-    end
-
-    it "returns correct sum [20 points]" do
-      expect(sum([1,2,3,4,5])).to be_a_kind_of Fixnum
+    it "returns correct sum" do
       expect(sum([1,2,3,4,5])).to eq(15)
-      expect(sum([1,2,3,4,-5])).to eq(5)
-      expect(sum([1,2,3,4,-5,5,-100])).to eq(-90)
     end
-
-    it "works on the empty array [10 points]" do
-      expect { sum([]) }.not_to raise_error
+    it "works on the empty array" do
       expect(sum([])).to be_zero
     end    
   end
 
   describe "#max_2_sum" do
-    it "should be defined" do
-      expect { max_2_sum([1,2,3]) }.not_to raise_error
-    end
-    it "returns the correct sum [7 points]" do
-      expect(max_2_sum([1,2,3,4,5])).to be_a_kind_of Fixnum
+    it "returns the correct sum" do
       expect(max_2_sum([1,-2,-3,-4,-5])).to eq(-1)
     end
-    it 'works even if 2 largest values are the same [3 points]' do
+    it 'works even if 2 largest values are the same' do
       expect(max_2_sum([1,2,3,3])).to eq(6)
     end
-    it "returns zero if array is empty [10 points]" do
+    it "returns zero if array is empty" do
       expect(max_2_sum([])).to be_zero
     end
-    it "returns value of the element if just one element [10 points]" do
+    it "returns value of the element if just one element" do
       expect(max_2_sum([3])).to eq(3)
     end
   end
 
   describe "#sum_to_n" do
-    it "should be defined" do
-      expect { sum_to_n?([1,2,3],4) }.not_to raise_error
-    end
-    it "returns true when any two elements sum to the second argument [30 points]" do
-      expect(sum_to_n?([1,2,3,4,5], 5)).to be true
+    it "returns true when any two elements sum to the second argument" do
       expect(sum_to_n?([3,0,5], 5)).to be true
-      expect(sum_to_n?([-1,-2,3,4,5,-8], 12)).to be false
-      expect(sum_to_n?([-1,-2,3,4,6,-8], 12)).to be false
     end
     
-    #for rspec 3.3.0
-    it "returns false for the single element array [5 points]" do
-      sum_to_n?([1], 1).should be_falsey
-      sum_to_n?([3], 0).should be_falsey
+    it "returns false for the single element array" do
+      expect(sum_to_n?([1], 1)).to be_falsey
     end
-    it "returns false for the empty array [5 points]" do
-      sum_to_n?([], 0).should be_falsey
-      sum_to_n?([], 7).should be_falsey
+    it "returns false for the empty array" do
+      expect(sum_to_n?([], 0)).to be_falsey
     end
-    #for rspec 2.14.1
-    # it "returns false for the single element array [5 points]" do
-    #   expect(sum_to_n?([1], 1)).to be false
-    #   expect(sum_to_n?([3], 0)).to be false
-    # end
-    # it "returns false for the empty array [5 points]" do
-    #   expect(sum_to_n?([], 0)).to be false
-    #   expect(sum_to_n?([], 7)).to be false
-    # end
   end
 end
