+++ spec/part2_spec.rb
@@ -1,3 +1,4 @@
+require 'ruby_intro.rb'
 describe "#hello" do
   it "should be defined" do
     expect { hello("Testing") }.not_to raise_error()#::NoMethodError)
