+++ spec/part3_spec.rb
@@ -1,3 +1,4 @@
+require 'ruby_intro.rb'
 describe "BookInStock" do
   it "should be defined" do
     expect { BookInStock }.not_to raise_error
